Fix flickering hover for stats puddles.
The stats annotations would flicker when moving the mouse around
inside the puddle. This was caused because the puddle is composed
of distint SVG elements inside a g element, and SVG g elements
don't fire hover events on there own. So I added a hidden rectangle
to trigger the events.